pyproject.toml
[project]
name = "livekit-avatar-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bithuman>=0.8.1",
    "gtts>=2.5.4",
    "livekit>=1.0.23",
    "livekit-agents[images]>=1.3.12",
    "livekit-api>=1.1.0",
    "livekit-plugins-anam>=1.3.12",
    "livekit-plugins-bithuman==1.3.12",
    "livekit-plugins-google>=1.3.12",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.3.12",
    "livekit-plugins-simli>=1.3.12",
    "livekit-plugins-tavus>=1.3.12",
    "livekit-plugins-turn-detector>=1.3.12",
    "loguru>=0.7.3",
    "numpy>=1.24,<2.4",
    "opencv-python>=4.13.0.90",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
